Update print-n-move to fix bug with bad text
when we hit a bad text file, we weren't moving it out of the 'fresh' folder, so we'd get stuck forever.

Now, print-n-move prints the name of the file it's trying to print, plus it moves the file to 'old' before trying to print, in case we error out during the print phase.